Revert "Use Balances API rather than BalanceByAddress"
PR https://github.com/penumbra-zone/wallet/pull/13 targeted main branch,
when it should have targeted the dev branch, according to git-flow
used for this repo. Undoing the change into main, so main stays pristine
and tracks the stable release in the Web Store.

We'll need to follow up and recommit on dev branch.

This reverts commit 915b18572ba6297e0f053a092d2ef9b28576ddd0.